+++ python/tvm/relay/op/strategy/cuda.py
@@ -1397,6 +1397,7 @@ def dft_strategy_cuda(attrs, inputs, out_type, target):
     )
     return strategy
 
+
 @layout_transform_strategy.register(["cuda", "gpu"])
 def layout_transform_strategy(attrs, inputs, out_type, target):
     strategy = _op.OpStrategy()
@@ -1406,287 +1407,3 @@ def layout_transform_strategy(attrs, inputs, out_type, target):
         name="layout_transform.cuda",
     )
     return strategy
-
-
-import tvm
-
-@tvm.register_func("meta_schedule.cuda.layout_transform")
-def cuda_layout_transform_schedule_rule(sch, block):
-    # params: input_buffer, output_buffer
-    params = sch.mod["main"].params
-    input_buffer = sch.mod["main"].buffer_map[params[0]]
-    output_buffer = sch.mod["main"].buffer_map[params[1]]
-
-    input_shape = [int(dim) for dim in input_buffer.shape]
-    output_shape = [int(dim) for dim in output_buffer.shape]
-
-    src_layout = sch.get_sref(block).stmt.annotations["src_layout"]
-    dst_layout = sch.get_sref(block).stmt.annotations["dst_layout"]
-
-    import math
-    from typing import List, Sequence, Tuple
-
-    from tvm.tir.schedule import BlockRV, ExprRV, LoopRV
-
-    def schedule_layout_transform_v4(
-        sch: tvm.tir.Schedule,
-        block_write: BlockRV,
-        src_layout: str,
-        dst_layout: str,
-        input_shape: List[int],
-        tile_size: ExprRV,
-    ):
-        ## Tiling block_read
-        # Let N and M represent the dimensions of interest
-        # N and M are the last dim of src_layout and dst_layout respectively.
-        # Then the initial block's loop will look like
-        # [i1, i2 ... M ... j1, j2 ... N]
-        #
-        # To guarantee contiguous read for N, we must group reads
-        # so that loops which factor N, j_n, j_n-1 ... are the innermost dimension
-        # Therefore our strategy for guaranteeing contiguous writes for N is by
-        # continually splitting inner-most dimension in order of N, j_n, j_n-1...
-        # by factors which divide into tile_size until we have the final tile_size.
-        # e.g. if tile_size = 32. N = 2, j_n = 2, j_n-1 = 4, j_n-2 = 24
-        # Then by combining N...j_n-1 we get a factor of 16, by spliiting up j_n-2 into
-        # two loops of 2 and 12, we can combine with the new loop of factor 2 to get a
-        # factor of 32. Note things don't divide evenly often time so we may have to pad
-        # to properly factorize.
-        #
-        # Similarly with M, to have contiguous writes we must consider the dst_layout:
-        # [a1, a2 ... N ... b1, b2 ... M]
-        # So that loops which factor M, b_m, b_m-1 ... are the innermost dimension
-        # Note that the dimension b_m, b_m-1 and j_n, j_n-1 may refer to the same dimension!
-        # By factoring j_n, j_n-1 we may be contributing to the innermost write dimension for
-        # M. However we note that both reads and writes must ideally have the same amount of
-        # work per thread for layout transforms so we still must build out to factors up
-        # to tile_size.
-
-        def pad_dimension_to_at_least_number(loop: LoopRV, requested_size: int):
-            """E.g. if loop has extant of 8 but we want 10, returns size 10 loop with padding"""
-            l1, l2 = sch.split(loop, [None, requested_size])
-            return sch.fuse(l1, l2)
-
-        def pad_dimension_to_factor_of_tile_size(
-            loop: LoopRV, initial_size: int, tile_size: int = tile_size
-        ) -> Tuple[LoopRV, int]:
-            """
-            Pads loop of given size until it is divisble into tile_size.
-            If the given size of the loop is greater than tile size. Do not pad.
-
-            example, loop_size = 5, tile_size = 32. loop_size --> 8
-                    loop_size = 5, tile_size = 36. loop_size --> 6
-                    loop_size = 8, tile_size = 32. loop_size --> 8
-                    loop_size = 33, tile_size = 32. loop_size --> 33
-
-            Returns padded loopRV and the new size
-            """
-            if tile_size % initial_size == 0:
-                return loop, int(initial_size)
-
-            if initial_size > tile_size or initial_size == tile_size:
-                return loop, int(initial_size)
-
-            # if initial_size > tile_size return without change, factor = 1
-            size = initial_size
-            while (tile_size % size) % tile_size > 0:
-                size += 1
-
-            return pad_dimension_to_at_least_number(loop, size), int(size)
-
-        def spin_out_factor(
-            loops: List[LoopRV], loop_extants: List[int], index: int, factor_needed: int
-        ) -> Tuple[List[LoopRV], List[int], int]:
-            """
-            Factor out loop extant to reach the requested factor. Updates the schedule in-place.
-
-            E.g. say we want to factors which eventually multiply to 32 (factor_needed).
-
-            Say we have the index we chose is a loop with an extant of 8.
-            E.g. loops / loop_extants = [3, 32, 6, 8], factor_needed = 32, index = 3
-                - 8 divides into 32 so we just split up the loop into two loops with extants 1 and 8.
-                - we then keep the 1-loop in place and move the new 8-loop to back of the list of loops
-                - ending loops / loop_extants = [3, 32, 6, 1, 8], remaining_factor_needed = 32 / 8 = 4
-
-            E.g. loops / loop_extants = [3, 32, 6, 8], factor_needed=32, index = 0
-                - 3 does not divide 32, so we pad until the extant divides 32, e.g. 4
-                - we then split up the loop into extants 1 and 4, moving the 4 to the back
-                - ending loops / loop_extants = [1, 32, 6, 8, 4], remaining_factor_needed = 32 / 4 = 8
-
-            E.g. loops / loop_extants = [3, 32, 6, 8], factor_needed=5, index = 3
-                - 8 is larger than 5 so we immediately do the splitting routine.
-                - the 8 extant loop becomes loops with extants 2 and 5
-                - ending loops / loop_extants = [1, 32, 6, 2, 5], remaining_factor_needed = 5 / 5 = 1
-
-            After updating loop ordering in place, returns the new list of loops, extants, and the
-            remaining factor needed.
-            """
-            cur_loop = loops[index]
-            cur_extant = loop_extants[index]
-
-            # Pad loops to divide evenly for factors needed, and split
-            new_loop, new_size = pad_dimension_to_factor_of_tile_size(
-                cur_loop, cur_extant, tile_size=factor_needed
-            )
-
-            split_factor = min(new_size, factor_needed)
-            new_loop_split, factored_loop = sch.split(new_loop, [None, split_factor])
-            factor_needed = factor_needed // split_factor
-
-            # update caching
-            loops[index] = new_loop_split
-            loops.append(factored_loop)
-
-            loop_extants[index] = math.ceil(new_size / split_factor)
-            loop_extants.append(split_factor)
-
-            sch.reorder(*loops)
-            return loops, loop_extants, factor_needed
-
-        def factor_dim_in_order(
-            indices: Sequence[int],
-            loops: List[LoopRV],
-            cur_loop_extants: List[int],
-            work_needed_inner_loop: int = tile_size,
-        ):
-            """TODO"""
-            for i in indices:
-                loops, cur_loop_extants, work_needed_inner_loop = spin_out_factor(
-                    loops, cur_loop_extants, i, work_needed_inner_loop
-                )
-                if work_needed_inner_loop == 1:
-                    break
-            return loops, cur_loop_extants
-
-        def get_high_level_loop_structure(block):
-            """Runs the factorization described above."""
-            # index 0 ... rank - 1 will always correspond to original loops
-            # perhaps after they have been factored.
-            loops = sch.get_loops(block)
-            cur_loop_extants = list(input_shape)
-
-            # Factor dim0 tile size and fuse things together
-            loops, cur_loop_extants = factor_dim_in_order(
-                range(rank - 1, -1, -1),
-                loops,
-                cur_loop_extants,
-                work_needed_inner_loop=tile_size,
-            )
-            # The factors which multiply to tile_size are now in back of our
-            # list of loops. However because we added them by traversing the inner
-            # dimensions, they are actually reversed order to guarantee the best access
-            # so reorder so reorder before fusing.
-            loops = loops[:rank] + loops[rank:][::-1]
-            cur_loop_extants = cur_loop_extants[:rank] + cur_loop_extants[rank::-1]
-            sch.reorder(*loops)
-            dim0_loop_tiled = sch.fuse(*loops[rank:])
-            loops = loops[:rank]
-            loops.append(dim0_loop_tiled)
-            cur_loop_extants = cur_loop_extants[:rank]
-            cur_loop_extants.append(tile_size)
-
-            # Same thing with dim1
-            # [:rank + 1], since we placed dim0_loop_tiled in the end which we want to keep
-            loops, cur_loop_extants = factor_dim_in_order(
-                (
-                    src_layout.index(dst_layout[loop_index_dst])
-                    for loop_index_dst in range(rank - 1, -1, -1)
-                ),
-                loops,
-                cur_loop_extants,
-                work_needed_inner_loop=tile_size,
-            )
-            loops = loops[: rank + 1] + loops[rank + 1 :][::-1]
-            cur_loop_extants = cur_loop_extants[: rank + 1] + cur_loop_extants[rank + 1 :: -1]
-            sch.reorder(*loops)
-            dim1_loop_tiled = sch.fuse(*loops[rank + 1 :])
-            loops = loops[: rank + 1]
-            loops.append(dim1_loop_tiled)
-            cur_loop_extants = cur_loop_extants[: rank + 1]
-            cur_loop_extants.append(tile_size)
-
-        rank = len(src_layout)
-
-        # Outer loop structure of read block matches that of src_layout
-        # E.g. if input_shape is [4, 6, 8]. Loops for read block will be
-        # for i, j, k in T.grid(4, 6, 8):
-        #     ...
-        # Read block will read from global memory coalesced at the start
-        # Assume write to output global memory is coalesced in block_write
-        block_read = sch.cache_read(block_write, 0, "shared")
-
-        # Here we have [loop1, loop2, loop3 ... dim0_tiled, dim1_tiled]
-        get_high_level_loop_structure(block_read)
-
-        loops = sch.get_loops(block_read)
-
-        # If there are insufficient elements, than dim1_tiled or dim0_tiled might be too small
-        # In all likelihood you should use a smaller tile, but I don't want things to crash.
-        loops[-1] = pad_dimension_to_at_least_number(loops[-1], tile_size)
-        loops[-2] = pad_dimension_to_at_least_number(loops[-2], tile_size)
-
-        # We want the dim0 and dim1 parent loops to be the inner most. Right now dim1 is inner-msot
-        # and we just need to move dim0 in (last dimension of dst).
-        # Recall right now structure is at least [l1 l2 ... ln, dim0_tiled, dim1_tiled]
-        # where n >= 2.
-        dim0_loop_index = src_layout.index(dst_layout[-1])
-        dim0_loop = loops.pop(dim0_loop_index)
-        loops = loops[:-3] + [dim0_loop, loops[-3]] + loops[-2:]
-        sch.reorder(*loops)
-
-        # After this: [outer_loop (block), dim0_tiled, dim1_tiled]
-        outer_loop = sch.fuse(*loops[:-2])
-
-        # Now that we have the high level loop structure, we can use reverse_compute_at magic
-        # To get the proper loop structure for writing! This is also as coalesced as possible
-        # already.
-        sch.reverse_compute_at(block_write, outer_loop)
-
-        # Fuse all inner loops for the write into 2 loops, grab inner loops for both read
-        # and write block which have locality (we will bind these to threadIdx)
-        fused_write_loop = sch.fuse(*sch.get_loops(block_write)[1:])
-        _, inner_write_loop = sch.split(fused_write_loop, [None, tile_size])
-        inner_read_loop = sch.get_loops(block_read)[-2]
-
-        sch.bind(loop=outer_loop, thread_axis="blockIdx.x")
-        sch.bind(loop=inner_write_loop, thread_axis="threadIdx.x")
-        sch.bind(loop=inner_read_loop, thread_axis="threadIdx.x")
-
-    from collections import deque
-    def auto_inline(start_block):
-        # BFS from start block in a chain (no branches)
-        fringe = deque([start_block])
-        visited = set()
-        while len(fringe) > 0:
-            cur_block = fringe.popleft()
-            if cur_block in visited:
-                continue 
-            else:
-                visited.add(cur_block)
-                
-            consumer_blocks = sch.get_consumers(cur_block)
-            
-            if len(consumer_blocks) >= 1:
-                fringe.extend(consumer_blocks)
-                sch.compute_inline(cur_block)
-            else:
-                # consumer yay!
-                return cur_block
-
-    schedules = []
-    
-    # For each schedule we also want to inline each stage as would be done in normal circumstances
-    # The block which producers the layout transform block seems 
-    block = auto_inline(block)
-
-    # Tile size 2,3,4...64
-    # Tile size of 1 does not make sense...
-    for tile_size in range(2, 65):
-        cur_sch = sch.copy()
-        schedule_layout_transform_v4(cur_sch, block, src_layout, dst_layout, input_shape, tile_size)
-        schedules.append(cur_sch)
-
-    # Also include the default schedules which will be handled via AutoBind schedule rule
-    schedules.append(sch)
-        
-    return schedules
