+++ python/tvm/relay/op/strategy/generic.py
@@ -21,8 +21,7 @@ import re
 
 from tvm import _ffi, ir, te, topi
 from tvm.target import generic_func, override_native_generic_func
-from tvm.topi.utils import (get_const_float, get_const_int, get_const_tuple,
-                            get_float_tuple)
+from tvm.topi.utils import get_const_float, get_const_int, get_const_tuple, get_float_tuple
 
 from .. import op as _op
 
@@ -2071,7 +2070,7 @@ def layout_transform_strategy(attrs, inputs, out_type, target):
 
 
 def wrap_compute_layout_transform(topi_compute, schedule_rule=""):
-    """Wrap stft compute"""
+    """Wrap layout transform compute"""
 
     def _compute_layout_transform(attrs, inputs, output_type):
         return [
