--- python/tvm/topi/x86/new.py
@@ -1,12 +1,0 @@
-import tensorflow as tf
-
-indices = tf.constant([[1], [1]])
-updates = tf.constant(
-    [
-        [[5, 5, 5, 5], [6, 6, 6, 6], [7, 7, 7, 7], [8, 8, 8, 8]],
-        [[5, 5, 5, 5], [6, 6, 6, 6], [7, 7, 7, 7], [8, 8, 8, 8]],
-    ]
-)
-shape = tf.constant([4, 4, 4])
-scatter = tf.scatter_nd(indices, updates, shape)
-print(scatter)
